# Optimize working with EllipseGeometry/RectangleGeometry, reduce allocs