## Description

Removes some unnecessary `gen0` allocations when working with `EllipseGeometry`/`RectangleGeometry` (such as adding them to a `PathGeometry` or combining geometries which serializes them (`Point` array alloc) and when rendering `TickBar`. Also introduces `internal` constructors for `PathFigure`/`PathGeometry` that can be used with inline arrays.

Also changes some `private static` fields to constants, decreasing static allocations.

The savings are not that massive right now, but it will allow for further optimizations (and there's a lot of space for it).

### Define geometry, add to group, get bounds

| Method       | Mean [ns]  | Error [ns] | StdDev [ns] | Gen0   | Code Size [B] | Allocated [B] |
|---------------- |-----------:|-----------:|------------:|-------:|--------------:|--------------:|
| Original | 9,371.1 ns |   67.37 ns |    56.25 ns | 0.4578 |       9,483 B |        7848 B |
| PR_EDIT | 8,986.0 ns |   55.25 ns |    48.98 ns |  0.4120    |   9,120 B  |        7120 B |

<details>
<summary>Benchmark code</summary>

```csharp
public Rect Original()
{
  PathGeometry abcd =  new PathGeometry();
  RectangleGeometry rectangle =   new RectangleGeometry(s_myRect, 10, 10);
  EllipseGeometry ellipse = new EllipseGeometry(s_myPoint, 10, 10);

  // uses GetPathFigureCollection
  abcd.AddGeometry(rectangle);
  abcd.AddGeometry(ellipse);

  // Retrieve bounds, another alloc
  Rect myBounds = rectangle.Bounds;
  Rect myEllipse = ellipse.Bounds;

  myBounds.Union(myEllipse);

  return myBounds;
}
```

</details>


## Customer Impact

Improved performance, decreased allocations.

## Regression

No.

## Testing

Local build, sample apps, usage of types.

## Risk

Should be low, all the changes were done with using stack-allocated memory