@h3xds1nz 2024-10-23
@robert-abeo I wasn't originally even gonna respond, but here you go.

1) The team is free to close my PRs if they deem them unfit, I'm sure they do not need your permission.
2) There's a saying you probably are not familiar with (this time, not your fault); it goes roughly as "he opened a restaurant but people started coming in". It's a nice but a very true oxymoron, much of like what you've said here.

3) You must be the reason x86_64 architecture still starts up in real mode. Every bug fix is a risk, every upstream level change (JIT, BCL) risks breaking all downstream consumers (e.g production apps) for, as you say, no benefit. Why even introduce new types, why even introduce ref structs, ability to `stackalloc` memory, and actually using those new features under the hood? I'd start at the core, that's dotnet/runtime. For some reason, they wanted `NET` unlike `NetFX` to be performance driven. You should go ahead and create an issue to tell everyone how you feel about the fact that about 70% of the BCL has already been rewritten using new language and runtime features for often low change-to-change benefit but overall great results, since most changes are often a building block for the next one.
Most optimizations are incremental, sometimes you need to shave off 1 kB out of 8 kB so then you can get additional 5 kB down and 60% perf improvement in the next; not everytime you get x7 perf improvement and 10% of previously allocated memory the first time around. But why do I explain it to you? Besides everyone else's, I'm just breaking the platform my company builds apps on and about 40% of the infrastructure depends on.

4) I've resolved your conversation because it has made zero sense. The WPF codebase has obviously had changing style over the years, some parts are written with pure C/C++ conventions (member names, function names, etc.) even though they do not incorporate or relate to any unmanaged calls simply because developers writing them were used to ATL/MFC code style, some parts are generated with that same conventions; some follow early `NetFX` style (namings etc.), then there are parts written after the release of C# 3.0 that incorporate `var` (there are few but there are some) which is generally discouraged (nobody wants to think about JavaScript), and so much more variations that I could spend days on.

There are sometimes `static` functions that start with `_` (hello C-style, I'd like you see you write this in a .NET project); members that have `m_` prefix; there are constants that have `c_` prefix, there are constants that have `_` prefix, and then there are finally constants in `PascalCase` as it is recommended by the guidelines these days.

But guess what, shocker, most of the time, the conventions vary not just in a namespace, but in a single file, and in the particular file (EllipseGeometry.cs) where you've highligted the line 72; I'd recommend checking line 360; 205; 157 (original lines, those were not edited since that part of WPF was opensourced), and tell me, please, what is the coding style? Why are those function definitions as long as the particular `ctor` I've auto-formatted (yes, an analyzer, imagine that). Guess opposite of so called coding style in one file. If anything, I've just unified it.

If we wanna have a look at so called "Microsoft's standard code format rules", guess those would be mostly incorporated in standard tools such as Visual Studio and Roslyn analyzers, correct? You obviously weren't thinking about incorporating random Rider rules and line breaks for 1024x768 CRT monitors, calling them current "Microsoft's standard code format rules"? Plus there's of course runtime coding style which I can't really seem to have violated. https://github.com/dotnet/runtime/blob/main/docs/coding-guidelines/coding-style.md

Not even in C code bases you'd see auto formatters mostly to wrap 1 per line. Plus, as of like 4 years ago, there's `--ignore-revs-file with` with `git blame` where you put those 2-3 big auto format commits and nobody gets hurt. So, for example:

- I've renamed constants (where some were defined fields and some `const`), both prefixed with `c_` as PascalCase
    - That would fall under `IDE1006`
- I've used pattern matching, that would be `IDE0019`
- I've inlined out variables, that would be C# 7? and `IDE0018`
- I've simplified some object inits into `new()` on right-hand side, that would be `IDE0090`
- I've simplified `PathGeometryData` init, that was `IDE0017`
- I've renamed some data types (e.g. `UInt32`) to language keywords, that's IDE0049
- I've simplified some collection initializations so that would be `IDE0300`
    - Aas that's the only way to get Roslyn to emit `InlineArray` parameters, not because I'm collection expressions fan
- I've changed compile-time constants that are defined as "static private byte" (not even `readonly` as it wasn't available at the time of writing) into actual constant definitions (`const`).
    - There is no compile-time checking without `readonly` so anyone could overwrite it in the future
    - It consumes unnecessary memory (perf opt)
    - JIT cannot treat them as constants
    - `static` first is mostly C-style, not really the newest C#
- I've changed `Point * name` (don't know what this is) and `Point *name` (C-Style) to `Point* name` - VS auto format
- I've added whitespaces between numbers on multiplication and between parentheses and code blocks
   - e.g. `fixed(byte` - VS auto format
- At line 384 and 390 (original lines) in `RectangleGeometry`, what is the coding style? Confused. Same issue as you highlighted above with `EllipseGeometry`. One is that way, second time it's the other way. This time it's an `if` condition.

And yes, it is absolutely fine to put parameters onto the next line when a single one gets too long!

Lastly, given the time and effort you must have put into your two comments, you could have put up PRs fixing half of your Fluent-styling issues instead and attempt to help everyone using WPF's new Fluent theme for a much better experience incorporating it into their applications. I'd even give you a thumbs up as it shows effort! But I understand, it is always easier to randomly criticise from distance. Please, instead of responding to my reaction here, I'd recommend creating a discussion, name it "Performance optimizations PRs should be banned as the team has not been 100% focused on the issues I have created" if you want and have a go at it. But I shall not take this from a drive-by shooter. Fixing community-reported bugs and alongside doing some optimizations does not make me the ultimate authority but it surely gives me more ground to have an educated opinion on this, and given your inability to even go through one file before saying how I've violated the codestyle that's violated 15 times in that file to begin with, it definitely doesn't hold a high value to me as a constructive feedback.

I couldn't be more constructive, simply because, well, I "focus on bigger gains" but without preparing the ground at the down level, with PRs like these, you can never get there. One last example, let's see if you can comprehend it: `Span<T>`.

With that being said, I'll continue on "spamming" when I feel like it unless the team's position is to close my PRs, thank you. There's 1.2k open issues and for about 80% of what I'm submitting it is contributing to fixing those issues but since it is only partially as the fixes require ground preparation often times, it does not justify slapping "Fixes #" on it.